Alterações do tipo da variável para a plataforma NodeMCU.

+++ src/DCMotorDriver.h
@@ -17,19 +17,19 @@
 class DCMotorDriver
 {
 	public:
-			DCMotorDriver(uint8_t enable, uint8_t input1, uint8_t input2);
-			void Speed(uint8_t speedPwm);
-			void Forward(uint8_t speed = 0);
-			void Backward(uint8_t speed = 0);
+			DCMotorDriver(int enable, int input1, int input2);
+			void Speed(int speedPwm);
+			void Forward(int speed = 0);
+			void Backward(int speed = 0);
 			void Stop();
 			
 	private:
-			void Move(uint8_t direction, uint8_t speed = 0);
-			uint8_t _en;
-			uint8_t _in1;
-			uint8_t _in2;
-			uint8_t _speedPwm;
-			uint8_t _direction;
+			void Move(int direction, int speed = 0);
+			int _en;
+			int _in1;
+			int _in2;
+			int _speedPwm;
+			int _direction;
 };
 
 #endif
